+++ je_auto_control/utils/subpixel_match/__init__.py
@@ -1,0 +1,6 @@
+"""Sub-pixel refinement of a template match by quadratic peak fitting."""
+from je_auto_control.utils.subpixel_match.subpixel_match import (
+    SubPixelMatch, match_subpixel, refine_peak,
+)
+
+__all__ = ["SubPixelMatch", "match_subpixel", "refine_peak"]